Separated RWS 1.0 and RWS 2.0 classes
