+++ src/rws_subscription.cpp
@@ -1,6 +1,5 @@
 #include <abb_librws/rws_subscription.h>
 #include <abb_librws/rws_error.h>
-#include <abb_librws/rws_client.h>
 #include <abb_librws/parsing.h>
 
 #include <Poco/Net/HTTPRequest.h>
